app/edge/requirements.txt
autopylogger==1.0.0
python-dotenv==1.0.0
pyserial==3.5.0
dataclasses-json==0.6.3
kafka-python==2.0.2
